make this code compile again with qreal != double architectures
svn path=/trunk/koffice/; revision=920891

+++ kplato/libs/ui/kptganttitemdelegate.cpp
@@ -427,7 +427,7 @@ void GanttItemDelegate::paintGanttItem( QPainter* painter, const KDGantt::StyleO
             const QRectF r = QRectF( itemRect ).adjusted( -pw, -pw, pw, pw );
             QPainterPath path;
             const qreal deltaY = r.height()/2.;
-            const qreal deltaX = qMin( r.width()/2., deltaY );
+            const qreal deltaX = qMin( r.width()/qreal(2), deltaY );
             path.moveTo( r.topLeft() );
             path.lineTo( r.topRight() );
             path.lineTo( QPointF( r.right(), r.top() + 2.*deltaY ) );
